Make sync.profile safer to evaluate.
This file was unconditionally running some external commands when
loaded, which is very dangerous.  The correct fix is to only put static
data into this file.  However, we are currently lacking a proper
configure test solution for Qt modules, which is needed for qtsystems.

As a temporary compromise, disable the hacky code in all cases except
when the file is being evaluated by syncqt.

Change-Id: I057778597a8ce69fca0bd0fd494d50b957d14956
Reviewed-on: http://codereview.qt.nokia.com/1855
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Andrew Stanley-Jones <andrew.stanley-jones@nokia.com>