Fix signalling order issues with blocking methods
If a method with a return value is called in the service, and this
method generates a series of signals, the client would process
these signals before the original function returned.  This leads
to confusion and problem synchronizing a client and service.

Change-Id: Idb80a3d4674b79904c5f96cadaeed7b3be939d35
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Andrew Stanley-Jones <andrew.stanley-jones@nokia.com>