Make certain that qServiceDebugLog can't block; clean up a little
It's possible the OS could be keeping buffers on the UDP sockets for
some reason and these will fill up, resulting in the debug logger
blocking. This patch detects this case by using the UDP socket in
O_NONBLOCK mode, and handles it by queuing up unsent messages for
later.

Also some cleanups that reduce the amount of code left behind by
qServiceDebug calls when CONFIG+=sfwdebug is _not_ set.

Change-Id: Iab50c013d59652a8a54bce2a2aba86789c398ec0
Reviewed-by: Andrew Stanley-Jones <andrew.stanley-jones@nokia.com>