Compile qnetworkinfo_linux.cpp under C++11 mode
Our c++11.prf file puts the compiler in standards-compliant mode
(-std=c++0x or -std=c++11), but unfortunately bluetooth.h is not
standards-compliant. It makes use of typeof(), which is a GCC
extension, without the double double-underscore marker (__extension__
and __typeof__).

To support this until upstream fixes their mess, put the compiler in
C++11-plus-extensions mode.

Change-Id: I5e0925ce90fb6564db0ab7b096d1967a1d3e3df4
Reviewed-by: Lincoln Ramsay <lincoln.ramsay@nokia.com>