Fix unit tests, remove race condition
Service registration should only be done in 1 place.  For unit tests
this is in the test client.  Services should not register/unregister
since they don't know when the test starts or stops, and can
unregister services in the middle of test runs.  Don't do this.

Also fixes the client to fail more quickly if errors are
detected when creating threads.

Change-Id: I51c351bf2c7c2cc83da604dfde27443840d02fb8
Reviewed-by: Lincoln Ramsay <lincoln.ramsay@nokia.com>