Do not use interface as variable or parameter name
As "interface" is defined as struct in objbase.h on windows it either
should not be used as a variable or parameter name or undefed before.
(see CI failure in https://codereview.qt-project.org/#change,45321 )

As parts of QtSystems rely on other windows headers which need interface
defined as struct it cannot be 'undef'ed "globally". Thus "interface" is
replaced with proper values in files where it occurs.

Change-Id: I9f77df966c7243fbc06aea87be075bfbd9b05d4a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Rainer Keller <rainer.keller@digia.com>
Reviewed-by: Cristiano di Flora <cristiano.diflora@gmail.com>