make x11 dependency a qmake CONFIG opt-out
This make it possible to compile with x11 in the build system, but not have
the x11 dependency. The tradeoff is that QScreenSaver disabling/enabling of
screenblanking is not possible at this time.

It may be possible using wayland, if not now sometime in the future.
In that case, QScreenSaver will be made to work with that.

Change-Id: I9f230dc906abc6eeafe2a18d31014974e8903442
Reviewed-by: Alex <alexander.blasche@digia.com>