QUPowerDeviceInterface: Make every property request except the first one asynchronous.
Fetching UPower properties can take a long time (I've seen >20ms on a device
that isn't even under load), so minimizing the amount of time we wait on the
main thread for the reply makes sense.

We keep the properties we had before around, and only emit changed signals when
the properties actually change.

Change-Id: Ia3a111d16bf0149a9aa153bf08f37409663ff7be
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>